Use testKeys to test different behavior on the confirmation page
